Music Genre Classifier
The overall goal of the project was to create a machine learning model capable of classifying music genres based on the acoustic features extracted from audio files. The project showcased common practices in audio signal processing, deep learning model development, and the handling of multi-segment predictions for music data.
Here's a brief description of the key steps and components:

1. Dataset and Preprocessing:
The project started with a music dataset organized into genre subfolders.
The librosa library was used to extract Mel-Frequency Cepstral Coefficients (MFCCs) from audio files, which are commonly used features in audio analysis.
The dataset was divided into segments, and MFCCs were extracted for each segment.

2.Data Storage:
The extracted MFCCs, genre labels, and a mapping of genre labels to class indices were stored in a JSON file (data_10.json).

3.Model Training:
A deep learning model was built using TensorFlow and Keras.
The model consisted of a sequence of densely connected layers.
The model was trained using the preprocessed data, with a focus on classifying music genres based on the extracted MFCCs.

4.Model Evaluation:
The model's performance was evaluated using a validation set to ensure it learned patterns that generalize well to new, unseen data.
Key metrics such as accuracy and loss were monitored during training.
The accuracy crossed 95% whereas the validation accuracy was close to 70%.
Further addition of  dense layers and increased epochs could result in validation accuracy in the range of 80-85%. This model is built on open source music data available. Having a more dense and diverse data could definitely help in improving the performance of the model.

5.Making Predictions:
Once the model was trained, it was used to make predictions on new audio files.
For each song, predictions were made for 10 segments, and the final prediction for the entire song was obtained by taking the maximum probability across the segments.

6.Genre Label Mapping:
The project included a mapping of class indices to genre labels, enabling the interpretation of model predictions in human-readable terms.

7.Output Presentation:
The final predictions, indicating the predicted genre for each song, were presented in a readable format.